+++ includes/router.php
@@ -113,6 +113,9 @@ function ms_handle_app_routes(array $config): bool
     if (ms_handle_config_page($config)) {
         return true;
     }
+    if (ms_handle_toolkit_page($config)) {
+        return true;
+    }
 
     return false;
 }